+++ modules/control/test.py
@@ -112,7 +112,7 @@ def test_adapters(prompt, negative, image):
             model_id = 'None'
         if shared.state.interrupted:
             continue
-        output = image
+        output = image.copy()
         if model_id != 'None':
             adapter = adapters.Adapter(model_id=model_id, device=devices.device, dtype=devices.dtype)
             if adapter is None:
@@ -122,6 +122,7 @@ def test_adapters(prompt, negative, image):
             pipe = adapters.AdapterPipeline(adapter=adapter.model, pipeline=shared.sd_model)
             pipe.pipeline.to(device=devices.device, dtype=devices.dtype)
             sd_models.set_diffuser_options(pipe)
+            image = image.convert('L') if 'Canny' in model_id or 'Sketch' in model_id else image.convert('RGB')
             try:
                 res = pipe.pipeline(prompt=prompt, negative_prompt=negative, image=image, num_inference_steps=10, output_type='pil')
                 output = res.images[0]
@@ -172,6 +173,59 @@ def test_xs(prompt, negative, image):
                 continue
             shared.log.info(f'Testing ControlNet-XS: {model_id}')
             pipe = controlnetsxs.ControlNetXSPipeline(controlnet=xs.model, pipeline=shared.sd_model)
+            pipe.pipeline.to(device=devices.device, dtype=devices.dtype)
+            sd_models.set_diffuser_options(pipe)
+            try:
+                res = pipe.pipeline(prompt=prompt, negative_prompt=negative, image=image, num_inference_steps=10, output_type='pil')
+                output = res.images[0]
+            except Exception as e:
+                errors.display(e, f'ControlNet-XS {model_id} inference')
+                model_id = f'{model_id} error'
+            pipe.restore()
+        draw = ImageDraw.Draw(output)
+        font = ImageFont.truetype('DejaVuSansMono', 48)
+        draw.text((10, 10), model_id, (0,0,0), font=font)
+        draw.text((8, 8), model_id, (255,255,255), font=font)
+        images.append(output)
+        yield output, None, None, images
+    rows = round(math.sqrt(len(images)))
+    cols = math.ceil(len(images) / rows)
+    w, h = 256, 256
+    size = (cols * w + cols, rows * h + rows)
+    grid = Image.new('RGB', size=size, color='black')
+    shared.log.info(f'Test ControlNet-XS: images={len(images)} grid={grid}')
+    for i, image in enumerate(images):
+        x = (i % cols * w) + (i % cols)
+        y = (i // cols * h) + (i // cols)
+        thumb = image.copy().convert('RGB')
+        thumb.thumbnail((w, h), Image.Resampling.HAMMING)
+        grid.paste(thumb, box=(x, y))
+    yield None, grid, None, images
+    return None, grid, None, images # preview_process, output_image, output_video, output_gallery
+
+
+def test_lite(prompt, negative, image):
+    from modules import devices, sd_models
+    from modules.control import controlnetslite
+    if image is None:
+        shared.log.error('Image not loaded')
+        return None, None, None
+    from PIL import ImageDraw, ImageFont
+    images = []
+    for model_id in controlnetslite.list_models():
+        if model_id is None:
+            model_id = 'None'
+        if shared.state.interrupted:
+            continue
+        output = image
+        if model_id != 'None':
+            lite = controlnetslite.ControlLLLite(model_id=model_id, device=devices.device, dtype=devices.dtype)
+            if lite is None:
+                shared.log.error(f'Control-LLite load failed: id="{model_id}"')
+                continue
+            shared.log.info(f'Testing ControlNet-XS: {model_id}')
+            pipe = controlnetslite.ControlLLitePipeline(pipeline=shared.sd_model)
+            pipe.apply(controlnet=lite.model, image=image, conditioning=1.0)
             pipe.pipeline.to(device=devices.device, dtype=devices.dtype)
             sd_models.set_diffuser_options(pipe)
             try: