+++ modules/generation_parameters_copypaste.py
@@ -125,7 +125,8 @@ def bind_buttons(buttons, send_image, send_generate_info):
     for tabname, button in buttons.items():
         source_text_component = send_generate_info if isinstance(send_generate_info, gr.components.Component) else None
         source_tabname = send_generate_info if isinstance(send_generate_info, str) else None
-        register_paste_params_button(ParamBinding(paste_button=button, tabname=tabname, source_text_component=source_text_component, source_image_component=send_image, source_tabname=source_tabname))
+        bindings = ParamBinding(paste_button=button, tabname=tabname, source_text_component=source_text_component, source_image_component=send_image, source_tabname=source_tabname)
+        register_paste_params_button(bindings)
 
 
 def register_paste_params_button(binding: ParamBinding):